Better responsiveness, Edit layout button fix (crash when no layout selected), code refactoring, color scheme update
